Refactored App to start threads automatically
The `App` class now contains vector of pointer to `BaseThread` objects.
This vector is populated with the `App::RegisterThread()` function.
Since the `App` class holds a vector of pointers, the caller of
`RegisterThread` must ensure the lifetime of the thread is at least as
long as the lifetime of the app. In the example programs, this is
accomplished by having the threads as member variables on the
user-defined App object that derives from `cactus_rt::App`.

During `App::Start()`, the threads are started in registration order
automatically. No more need to manually create a `App::Start()`
function.

To enable this refactor, a `BaseThread` class has to be created as the
current `Thread` class is a templated class (and thus cannot be directly
used as the value type of a vector). Additionally, the `BaseThread` now
implements functionality to request the thread to stop via an atomic
stop flag so this code can be shared across all thread implementations.
`CyclicThread` now checks this flag via the `BaseThread::StopRequested()`.
We also use the `memory_order_relaxed` to check, as this is safe and
slightly more efficient than using sequential consistency (default for
atomics).

I've also considered to implement an `App::RequestStop()` function, but
stopping order could be a problem so I've deferred that for later.