Added iceoryx example build
Also had to rename the cmake variable `CLANG_TIDY` to `CLANG_TIDY_PATH`
because iceoryx expects a variable called `CLANG_TIDY` to enable its own
clang-tidy process. If we didn't rename this variable, it would mean
that `ENABLE_CLANG_TIDY=ON` would also enable clang-tidy for iceoryx,
which we don't want to do.