Added an example that communicate with iceoryx
Iceoryx is an interesting library that allows one to easily communicate
between processes and threads with zero latency using shared memory
approach. This example creates a RT application with two threads that
communicate with single-process iceoryx communication (as opposed to
multi-process, shared-memory-based communication).

If this is indeed a promising approach, perhaps it is a good idea to
incorporate iceoryx directly into the cactus-rt framework.