Explicitly instantiate CyclicThread<SchedulerT>
This is needed to move the code definition out of the header file.
Later, when we implement built-in tracing via Perfetto, we can't include
the Perfetto header into the Cactus RT headers. This is because defining
Perfetto tracing categories requires a pile of macros that won't play
very well when included and called multiple times in the same
translation unit (although this is resolved with [this commit][1], which
is not released at this time). Since the cyclic thread code need to
include the Perfetto headers, and user code depends on cyclic_thread.h,
the only way to sanely include headers is to move the cyclic thread code
into its own translation unit, which then requires explicit
instantiation of the different threads.

[1]: https://github.com/google/perfetto/commit/6020a92b81bb64bf66b99cf85d37eab2f8ded9fb