Fixed clangd in ROS workspace.
Adding a .clangd file in the root of the repo causes it to be used even
if cactus-rt is developed as a part of a ROS package, which causes
clangd to not show the appropriate completion data. By moving it into a
vscode config, we get around this problem for now.