Added basic scaffolding for ROS2 subscriber
This is pretty inefficiently implemented for now. This only implements
the latest value subscriber.

Need to significantly overhaul the code to not use a functional
converter, but a type which would enable better compile time detection
of wrong usage. Possibly reuse rclcpp::TypeAdapter for this. That's in a
future commit tho. Right now this sets up some temporary
code/scaffolding.

Will need to implement an "all value" subscriber with a queue as well,
but this will happen after the TypeAdapter overhaul. Some
simplification/optimization might also occur before that.