Real-time-safe random number generator.
Xorshift random number generator is a constant-time pseudorandom
generator that's real-time-safe.

We also create an "uniform" distribution function that converts the
random number generator (which generates a random 64 bit sequence) to a
number in [0, 1). This uniform distribution is guaranteed to be constant
time but is not 100% uniform as it does not regenerate a number if the
underlying generator generates its max value. In the case where the
underlying generator generates its max value, this distribution function
will generate 0.0. This means the chance of generating 0.0 is slightly
higher than every other value, making the distribution not truly
uniform. The non-uniformity should be very small with respect to the
Xorshift64 generator so it should work for non-cryptographic purposes
(which you probably shouldn't be doing in real-time anyway).