App::CreateThread
Instead of letting `Thread` be created outside of the app, we now force
the `Thread` to be created by the `App::CreateThread` factory method.
This allows us to better control lifetime and sets up the stage to
possibly eliminate the weak_ptr situation with the trace aggregator if
the Thread is not longer a shared_ptr.